refactor(test): Extract common crypto mock server helper
