+++ crates/matrix-sdk/tests/integration/encryption/to_device.rs
@@ -1,108 +1,21 @@
-use matrix_sdk::{
-    authentication::matrix::MatrixSession, config::RequestConfig,
-    test_utils::client::mock_session_tokens, Client,
-};
-use matrix_sdk_base::SessionMeta;
-use matrix_sdk_test::{async_test, test_json, SyncResponseBuilder};
-use ruma::{api::MatrixVersion, owned_device_id, owned_user_id, serde::Raw};
+use matrix_sdk::test_utils::mocks::encryption::MatrixKeysMockServer;
+use matrix_sdk_test::{async_test, test_json};
+use ruma::serde::Raw;
 use serde_json::json;
 use wiremock::{
     matchers::{method, path, path_regex},
     Mock, ResponseTemplate,
 };
 
-use crate::mock_sync_scoped;
-
-async fn set_up_alice_and_bob_for_encryption(
-    server: &mut crate::encryption::verification::MockedServer,
-) -> (Client, Client) {
-    let alice_user_id = owned_user_id!("@alice:example.org");
-    let alice_device_id = owned_device_id!("4L1C3");
-    let alice = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-    alice
-        .restore_session(MatrixSession {
-            meta: SessionMeta {
-                user_id: alice_user_id.clone(),
-                device_id: alice_device_id.clone(),
-            },
-            tokens: mock_session_tokens(),
-        })
-        .await
-        .unwrap();
-
-    let bob = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-
-    let bob_user_id = owned_user_id!("@bob:example.org");
-    let bob_device_id = owned_device_id!("B0B0B0B0B");
-    bob.restore_session(MatrixSession {
-        meta: SessionMeta { user_id: bob_user_id.clone(), device_id: bob_device_id.clone() },
-        tokens: mock_session_tokens(),
-    })
-    .await
-    .unwrap();
-
-    server.add_known_device(&alice_device_id);
-    server.add_known_device(&bob_device_id);
-
-    // Have Alice track Bob, so she queries his keys later.
-    {
-        let alice_olm = alice.olm_machine_for_testing().await;
-        let alice_olm = alice_olm.as_ref().unwrap();
-        alice_olm.update_tracked_users([bob_user_id.as_ref()]).await.unwrap();
-    }
-
-    // Have Alice and Bob upload their signed device keys.
-    {
-        let mut sync_response_builder = SyncResponseBuilder::new();
-        let response_body = sync_response_builder.build_json_sync_response();
-        let _scope = mock_sync_scoped(&server.server, response_body, None).await;
-
-        alice
-            .sync_once(Default::default())
-            .await
-            .expect("We should be able to sync with Alice so we upload the device keys");
-        bob.sync_once(Default::default()).await.unwrap();
-    }
-
-    // Run a sync so we do send outgoing requests, including the /keys/query for
-    // getting bob's identity.
-    let mut sync_response_builder = SyncResponseBuilder::new();
-
-    {
-        let _scope = mock_sync_scoped(
-            &server.server,
-            sync_response_builder.build_json_sync_response(),
-            None,
-        )
-        .await;
-        alice
-            .sync_once(Default::default())
-            .await
-            .expect("We should be able to sync so we get theinitial set of devices");
-    }
-
-    (alice, bob)
-}
 #[async_test]
 async fn test_encrypt_and_send_to_device() {
     // ===========
     // Happy path, will encrypt and send
     // ============
-    let mut server = crate::encryption::verification::MockedServer::new().await;
 
-    let (alice, bob) = set_up_alice_and_bob_for_encryption(&mut server).await;
+    let matrix_keys_server_mock = MatrixKeysMockServer::new().await;
+
+    let (alice, bob) = matrix_keys_server_mock.set_up_alice_and_bob_for_encryption().await;
     let bob_user_id = bob.user_id().unwrap();
     let bob_device_id = bob.device_id().unwrap();
 
@@ -134,7 +47,7 @@ async fn test_encrypt_and_send_to_device() {
         // Should be called once
         .expect(1)
         .named("send_to_device")
-        .mount(&server.server)
+        .mount(&matrix_keys_server_mock.server)
         .await;
 
     alice
@@ -149,9 +62,10 @@ async fn test_encrypt_and_send_to_device_report_failures_server() {
     // Error case, when the to-device fails to send
     // ============
 
-    let mut server = crate::encryption::verification::MockedServer::new().await;
+    let matrix_keys_server_mock = MatrixKeysMockServer::new().await;
 
-    let (alice, bob) = set_up_alice_and_bob_for_encryption(&mut server).await;
+    let (alice, bob) = matrix_keys_server_mock.set_up_alice_and_bob_for_encryption().await;
+
     let bob_user_id = bob.user_id().unwrap();
     let bob_device_id = bob.device_id().unwrap();
 
@@ -176,7 +90,7 @@ async fn test_encrypt_and_send_to_device_report_failures_server() {
         // There is retries in place, assert it
         .expect(3)
         .named("send_to_device")
-        .mount(&server.server)
+        .mount(&matrix_keys_server_mock.server)
         .await;
 
     let alice_bob_device = alice
@@ -203,9 +117,9 @@ async fn test_encrypt_and_send_to_device_report_failures_encryption_error() {
     // Error case, when the encryption fails
     // ============
 
-    let mut server = crate::encryption::verification::MockedServer::new().await;
+    let matrix_keys_server_mock = MatrixKeysMockServer::new().await;
 
-    let (alice, bob) = set_up_alice_and_bob_for_encryption(&mut server).await;
+    let (alice, bob) = matrix_keys_server_mock.set_up_alice_and_bob_for_encryption().await;
     let bob_user_id = bob.user_id().unwrap();
     let bob_device_id = bob.device_id().unwrap();
 
@@ -230,7 +144,7 @@ async fn test_encrypt_and_send_to_device_report_failures_encryption_error() {
         // Should be called once
         .expect(0)
         .named("send_to_device")
-        .mount(&server.server)
+        .mount(&matrix_keys_server_mock.server)
         .await;
 
     let alice_bob_device = alice
@@ -248,7 +162,7 @@ async fn test_encrypt_and_send_to_device_report_failures_encryption_error() {
         })))
         // Take priority
         .with_priority(1)
-        .mount(&server.server)
+        .mount(&matrix_keys_server_mock.server)
         .await;
 
     let result = alice
