+++ crates/matrix-sdk/tests/integration/encryption/verification.rs
@@ -1,401 +1,19 @@
-use std::{
-    collections::BTreeMap,
-    sync::{Arc, Mutex},
-};
-
 use assert_matches2::assert_matches;
 use futures_util::FutureExt;
-use imbl::HashSet;
 use matrix_sdk::{
-    authentication::matrix::MatrixSession,
-    config::RequestConfig,
     encryption::VerificationState,
-    test_utils::{client::mock_session_tokens, logged_in_client_with_server},
+    test_utils::{logged_in_client_with_server, mocks::encryption::MatrixKeysMockServer},
     Client,
 };
-use matrix_sdk_base::SessionMeta;
 use matrix_sdk_test::{async_test, SyncResponseBuilder};
-use ruma::{
-    api::{client::keys::upload_signatures::v3::SignedKeys, MatrixVersion},
-    encryption::{CrossSigningKey, DeviceKeys, OneTimeKey},
-    owned_device_id, owned_user_id,
-    serde::Raw,
-    user_id, CrossSigningKeyId, DeviceId, OneTimeKeyAlgorithm, OwnedDeviceId, OwnedOneTimeKeyId,
-    OwnedUserId,
-};
+use ruma::{owned_device_id, owned_user_id, user_id};
 use serde_json::json;
 use wiremock::{
     matchers::{body_json, method, path},
-    Mock, MockServer, Request, ResponseTemplate,
+    Mock, ResponseTemplate,
 };
 
 use crate::mock_sync_scoped;
-
-#[derive(Debug, Default)]
-struct Keys {
-    device: BTreeMap<OwnedUserId, BTreeMap<String, Raw<DeviceKeys>>>,
-    master: BTreeMap<OwnedUserId, Raw<CrossSigningKey>>,
-    self_signing: BTreeMap<OwnedUserId, Raw<CrossSigningKey>>,
-    user_signing: BTreeMap<OwnedUserId, Raw<CrossSigningKey>>,
-    // We only keep `OwnedDeviceId` as key, based on assumption that device ids will
-    // be unique per user (fine for this mock)
-    one_time_keys: BTreeMap<OwnedDeviceId, BTreeMap<OwnedOneTimeKeyId, Raw<OneTimeKey>>>,
-}
-
-impl Keys {
-    /// Mocks some endpoints associated to key queries and cross-signing.
-    async fn mock_endpoints(server: &MockServer, known_devices: Arc<Mutex<HashSet<String>>>) {
-        let keys = Arc::new(Mutex::new(Self::default()));
-
-        Mock::given(method("POST"))
-            .and(path("/_matrix/client/r0/keys/query"))
-            .respond_with(mock_keys_query(keys.clone()))
-            .mount(server)
-            .await;
-
-        Mock::given(method("POST"))
-            .and(path("/_matrix/client/r0/keys/upload"))
-            .respond_with(mock_keys_upload(known_devices.clone(), keys.clone()))
-            .mount(server)
-            .await;
-
-        Mock::given(method("POST"))
-            .and(path("/_matrix/client/unstable/keys/device_signing/upload"))
-            .respond_with(mock_keys_device_signing_upload(keys.clone()))
-            .mount(server)
-            .await;
-
-        Mock::given(method("POST"))
-            .and(path("/_matrix/client/unstable/keys/signatures/upload"))
-            .respond_with(mock_keys_signature_upload(keys.clone()))
-            .mount(server)
-            .await;
-
-        Mock::given(method("POST"))
-            .and(path("/_matrix/client/r0/keys/claim"))
-            .respond_with(mock_keys_claimed_request(keys.clone()))
-            .mount(server)
-            .await;
-    }
-}
-
-pub(crate) struct MockedServer {
-    pub(crate) server: MockServer,
-    known_devices: Arc<Mutex<HashSet<String>>>,
-}
-
-/// Intercepts a `/keys/query` request and mock its results as returned by an
-/// actual homeserver.
-///
-/// Supports filtering by user id, or no filters at all.
-fn mock_keys_query(keys: Arc<Mutex<Keys>>) -> impl Fn(&Request) -> ResponseTemplate {
-    move |req| {
-        #[derive(Debug, serde::Deserialize)]
-        struct Parameters {
-            device_keys: BTreeMap<OwnedUserId, Vec<OwnedDeviceId>>,
-        }
-
-        let params: Parameters = req.body_json().unwrap();
-
-        let keys = keys.lock().unwrap();
-        let mut device_keys = keys.device.clone();
-        if !params.device_keys.is_empty() {
-            device_keys.retain(|user, key_map| {
-                if let Some(devices) = params.device_keys.get(user) {
-                    if !devices.is_empty() {
-                        key_map.retain(|key_id, _json| {
-                            devices.iter().any(|device_id| &device_id.to_string() == key_id)
-                        });
-                    }
-                    true
-                } else {
-                    false
-                }
-            })
-        }
-
-        let master_keys = keys.master.clone();
-        let self_signing_keys = keys.self_signing.clone();
-        let user_signing_keys = keys.user_signing.clone();
-
-        ResponseTemplate::new(200).set_body_json(json!({
-            "device_keys": device_keys,
-            "master_keys": master_keys,
-            "self_signing_keys": self_signing_keys,
-            "user_signing_keys": user_signing_keys,
-        }))
-    }
-}
-
-/// Intercepts a `/keys/upload` query and mocks the behavior it would have on a
-/// real homeserver.
-///
-/// Inserts all the `DeviceKeys` into `Keys::device_keys`, or if already present
-/// in this mapping, only merge the signatures.
-fn mock_keys_upload(
-    known_devices: Arc<Mutex<HashSet<String>>>,
-    keys: Arc<Mutex<Keys>>,
-) -> impl Fn(&Request) -> ResponseTemplate {
-    move |req: &Request| {
-        #[derive(Debug, serde::Deserialize)]
-        struct Parameters {
-            device_keys: Option<Raw<DeviceKeys>>,
-            one_time_keys: Option<BTreeMap<OwnedOneTimeKeyId, Raw<OneTimeKey>>>,
-        }
-
-        let params: Parameters = req.body_json().unwrap();
-
-        if let Some(new_device_keys) = params.device_keys {
-            let new_device_keys = new_device_keys.deserialize().unwrap();
-
-            let known_devices = known_devices.lock().unwrap();
-            let key_id = new_device_keys.device_id.to_string();
-            if known_devices.contains(&key_id) {
-                let mut keys = keys.lock().unwrap();
-                let devices = keys.device.entry(new_device_keys.user_id.clone()).or_default();
-
-                // Either merge signatures if an entry is already present, or insert a new one.
-                if let Some(device_keys) = devices.get_mut(&key_id) {
-                    let mut existing = device_keys.deserialize().unwrap();
-
-                    // Merge signatures.
-                    for (uid, sigs) in existing.signatures.iter_mut() {
-                        if let Some(new_sigs) = new_device_keys.signatures.get(uid) {
-                            sigs.extend(new_sigs.clone());
-                        }
-                    }
-                    for (uid, sigs) in new_device_keys.signatures.iter() {
-                        if !existing.signatures.contains_key(uid) {
-                            existing.signatures.insert(uid.clone(), sigs.clone());
-                        }
-                    }
-
-                    *device_keys = Raw::new(&existing).unwrap();
-                } else {
-                    devices.insert(key_id, Raw::new(&new_device_keys).unwrap());
-                }
-            }
-        }
-
-        if let Some(otks) = params.one_time_keys {
-            let mut keys = keys.lock().unwrap();
-
-            // We need a trick to find out what userId|device this OTK is for.
-            // This is not part of the payload, a real server uses the access token(?)
-            // Let's look at the signatures to find out
-            for (key_id, raw_otk) in otks {
-                let otk = raw_otk.deserialize().unwrap();
-                match otk {
-                    OneTimeKey::SignedKey(signed_key) => {
-                        let device_id = signed_key
-                            .signatures
-                            .first_key_value()
-                            .unwrap()
-                            .1
-                            .keys()
-                            .next()
-                            .unwrap()
-                            .key_name()
-                            .to_owned();
-
-                        keys.one_time_keys.entry(device_id).or_default().insert(key_id, raw_otk);
-                    }
-                    OneTimeKey::Key(_) => {
-                        // Ignore this old algorithm,
-                    }
-                    _ => {}
-                }
-            }
-        }
-
-        ResponseTemplate::new(200).set_body_json(json!({
-            "one_time_key_counts": {
-                // "signed_curve25519": 0
-            }
-        }))
-    }
-}
-
-/// Mocks a `/keys/device_signing/upload` request for bootstrapping
-/// cross-signing.
-///
-/// Assumes (and asserts) all keys are updated at the same time.
-///
-/// Saves all the different cross-signing keys into their respective fields of
-/// `Keys`.
-fn mock_keys_device_signing_upload(
-    keys: Arc<Mutex<Keys>>,
-) -> impl Fn(&Request) -> ResponseTemplate {
-    move |req: &Request| {
-        // Accept all cross-signing setups by default.
-        #[derive(Debug, serde::Deserialize)]
-        struct Parameters {
-            master_key: Option<Raw<CrossSigningKey>>,
-            self_signing_key: Option<Raw<CrossSigningKey>>,
-            user_signing_key: Option<Raw<CrossSigningKey>>,
-        }
-
-        let params: Parameters = req.body_json().unwrap();
-        assert!(params.master_key.is_some());
-        assert!(params.self_signing_key.is_some());
-        assert!(params.user_signing_key.is_some());
-
-        let mut keys = keys.lock().unwrap();
-
-        if let Some(key) = params.master_key {
-            let deserialized = key.deserialize().unwrap();
-            let user_id = deserialized.user_id;
-            keys.master.insert(user_id, key);
-        }
-
-        if let Some(key) = params.self_signing_key {
-            let deserialized = key.deserialize().unwrap();
-            let user_id = deserialized.user_id;
-            keys.self_signing.insert(user_id, key);
-        }
-
-        if let Some(key) = params.user_signing_key {
-            let deserialized = key.deserialize().unwrap();
-            let user_id = deserialized.user_id;
-            keys.user_signing.insert(user_id, key);
-        }
-
-        ResponseTemplate::new(200).set_body_json(json!({}))
-    }
-}
-
-/// Mocks a `/keys/signatures/upload` request.
-///
-/// Supports merging signatures for master keys or devices keys.
-fn mock_keys_signature_upload(keys: Arc<Mutex<Keys>>) -> impl Fn(&Request) -> ResponseTemplate {
-    move |req: &Request| {
-        #[derive(Debug, serde::Deserialize)]
-        #[serde(transparent)]
-        struct Parameters(BTreeMap<OwnedUserId, SignedKeys>);
-
-        let params: Parameters = req.body_json().unwrap();
-
-        let mut keys = keys.lock().unwrap();
-
-        for (user, signed_keys) in params.0 {
-            for (key_id, raw_key) in signed_keys.iter() {
-                // Try to find a field in keys.master.
-                if let Some(existing_master_key) = keys.master.get_mut(&user) {
-                    let mut existing = existing_master_key.deserialize().unwrap();
-
-                    let target = CrossSigningKeyId::from_parts(
-                        ruma::SigningKeyAlgorithm::Ed25519,
-                        key_id.try_into().unwrap(),
-                    );
-
-                    if existing.keys.contains_key(&target) {
-                        let param: CrossSigningKey = serde_json::from_str(raw_key.get()).unwrap();
-
-                        for (uid, sigs) in existing.signatures.iter_mut() {
-                            if let Some(new_sigs) = param.signatures.get(uid) {
-                                sigs.extend(new_sigs.clone());
-                            }
-                        }
-                        for (uid, sigs) in param.signatures.iter() {
-                            if !existing.signatures.contains_key(uid) {
-                                existing.signatures.insert(uid.clone(), sigs.clone());
-                            }
-                        }
-
-                        // Update in map.
-                        *existing_master_key = Raw::new(&existing).unwrap();
-                        continue;
-                    }
-                }
-
-                // Otherwise, try to find a field in keys.device.
-                // Either merge signatures if an entry is already present, or insert a new
-                // entry.
-                let known_devices = keys.device.entry(user.clone()).or_default();
-                let device_keys = known_devices
-                    .get_mut(key_id)
-                    .expect("trying to add a signature for a missing key");
-
-                let param: DeviceKeys = serde_json::from_str(raw_key.get()).unwrap();
-
-                let mut existing: DeviceKeys = device_keys.deserialize().unwrap();
-
-                for (uid, sigs) in existing.signatures.iter_mut() {
-                    if let Some(new_sigs) = param.signatures.get(uid) {
-                        sigs.extend(new_sigs.clone());
-                    }
-                }
-                for (uid, sigs) in param.signatures.iter() {
-                    if !existing.signatures.contains_key(uid) {
-                        existing.signatures.insert(uid.clone(), sigs.clone());
-                    }
-                }
-
-                *device_keys = Raw::new(&existing).unwrap();
-            }
-        }
-
-        ResponseTemplate::new(200).set_body_json(json!({
-            "failures": {}
-        }))
-    }
-}
-
-fn mock_keys_claimed_request(keys: Arc<Mutex<Keys>>) -> impl Fn(&Request) -> ResponseTemplate {
-    move |req: &Request| {
-        // Accept all cross-signing setups by default.
-        #[derive(Debug, serde::Deserialize)]
-        struct Parameters {
-            one_time_keys: BTreeMap<OwnedUserId, BTreeMap<OwnedDeviceId, OneTimeKeyAlgorithm>>,
-        }
-
-        let params: Parameters = req.body_json().unwrap();
-
-        let mut keys = keys.lock().unwrap();
-        let known_otks = &mut keys.one_time_keys;
-
-        let mut found_one_time_keys: BTreeMap<
-            OwnedUserId,
-            BTreeMap<OwnedDeviceId, BTreeMap<OwnedOneTimeKeyId, Raw<OneTimeKey>>>,
-        > = BTreeMap::new();
-
-        for (user, requested_one_time_keys) in params.one_time_keys {
-            for device_id in requested_one_time_keys.keys() {
-                let device_id = device_id.clone();
-                let found_key = known_otks
-                    .entry(device_id.clone())
-                    // .or_default();
-                    .or_default()
-                    .pop_first();
-                if let Some((id, raw_otk)) = found_key {
-                    found_one_time_keys
-                        .entry(user.clone())
-                        .or_default()
-                        .entry(device_id.clone())
-                        .or_default()
-                        .insert(id, raw_otk.clone());
-                }
-            }
-        }
-
-        ResponseTemplate::new(200).set_body_json(json!({
-            "one_time_keys" : found_one_time_keys
-        }))
-    }
-}
-
-impl MockedServer {
-    pub(crate) async fn new() -> Self {
-        let server = MockServer::start().await;
-        let known_devices: Arc<Mutex<HashSet<String>>> = Default::default();
-        Keys::mock_endpoints(&server, known_devices.clone()).await;
-        Self { server, known_devices }
-    }
-
-    pub(crate) fn add_known_device(&mut self, device_id: &DeviceId) {
-        self.known_devices.lock().unwrap().insert(device_id.to_string());
-    }
-}
 
 async fn bootstrap_cross_signing(client: &Client) {
     client.encryption().bootstrap_cross_signing(None).await.unwrap();
@@ -406,30 +24,15 @@ async fn bootstrap_cross_signing(client: &Client) {
 
 #[async_test]
 async fn test_own_verification() {
-    let mut server = MockedServer::new().await;
+    let server = MatrixKeysMockServer::new().await;
 
     let user_id = owned_user_id!("@alice:example.org");
     let device_id = owned_device_id!("4L1C3");
-    let alice = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-    alice
-        .restore_session(MatrixSession {
-            meta: SessionMeta { user_id: user_id.clone(), device_id: device_id.clone() },
-            tokens: mock_session_tokens(),
-        })
-        .await
-        .unwrap();
+    let alice = server.create_client(&user_id, &device_id).await;
 
     // Subscribe to verification state updates
     let mut verification_state_subscriber = alice.encryption().verification_state();
     assert_eq!(alice.encryption().verification_state().get(), VerificationState::Unknown);
-
-    server.add_known_device(&device_id);
 
     // Have Alice bootstrap cross-signing.
     bootstrap_cross_signing(&alice).await;
@@ -490,30 +93,15 @@ async fn test_own_verification() {
 
 #[async_test]
 async fn test_reset_cross_signing_resets_verification() {
-    let mut server = MockedServer::new().await;
+    let server = MatrixKeysMockServer::new().await;
 
     let user_id = owned_user_id!("@alice:example.org");
     let device_id = owned_device_id!("4L1C3");
-    let alice = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-    alice
-        .restore_session(MatrixSession {
-            meta: SessionMeta { user_id: user_id.clone(), device_id: device_id.clone() },
-            tokens: mock_session_tokens(),
-        })
-        .await
-        .unwrap();
+    let alice = server.create_client(&user_id, &device_id).await;
 
     // Subscribe to verification state updates
     let mut verification_state_subscriber = alice.encryption().verification_state();
     assert_eq!(alice.encryption().verification_state().get(), VerificationState::Unknown);
-
-    server.add_known_device(&device_id);
 
     // Have Alice bootstrap cross-signing.
     bootstrap_cross_signing(&alice).await;
@@ -547,22 +135,7 @@ async fn test_reset_cross_signing_resets_verification() {
     assert_eq!(alice.encryption().verification_state().get(), VerificationState::Verified);
 
     let device_id = owned_device_id!("AliceDevice2");
-    let alice2 = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-    alice2
-        .restore_session(MatrixSession {
-            meta: SessionMeta { user_id: user_id.clone(), device_id: device_id.clone() },
-            tokens: mock_session_tokens(),
-        })
-        .await
-        .unwrap();
-
-    server.add_known_device(&device_id);
+    let alice2 = server.create_client(&user_id, &device_id).await;
 
     // Have Alice bootstrap cross-signing again, this time on her second device.
     bootstrap_cross_signing(&alice2).await;
@@ -587,47 +160,16 @@ async fn test_reset_cross_signing_resets_verification() {
 
 #[async_test]
 async fn test_unchecked_mutual_verification() {
-    let mut server = MockedServer::new().await;
+    let server = MatrixKeysMockServer::new().await;
 
-    let alice_user_id = owned_user_id!("@alice:example.org");
-    let alice_device_id = owned_device_id!("4L1C3");
-    let alice = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
-    alice
-        .restore_session(MatrixSession {
-            meta: SessionMeta {
-                user_id: alice_user_id.clone(),
-                device_id: alice_device_id.clone(),
-            },
-            tokens: mock_session_tokens(),
-        })
-        .await
-        .unwrap();
-
-    let bob = Client::builder()
-        .homeserver_url(server.server.uri())
-        .server_versions([MatrixVersion::V1_0])
-        .request_config(RequestConfig::new().disable_retry())
-        .build()
-        .await
-        .unwrap();
+    let user_id = owned_user_id!("@alice:example.org");
+    let device_id = owned_device_id!("4L1C3");
+    let alice = server.create_client(&user_id, &device_id).await;
 
     let bob_user_id = owned_user_id!("@bob:example.org");
     let bob_device_id = owned_device_id!("B0B0B0B0B");
-    bob.restore_session(MatrixSession {
-        meta: SessionMeta { user_id: bob_user_id.clone(), device_id: bob_device_id.clone() },
-        tokens: mock_session_tokens(),
-    })
-    .await
-    .unwrap();
 
-    server.add_known_device(&alice_device_id);
-    server.add_known_device(&bob_device_id);
+    let bob = server.create_client(&bob_user_id, &bob_device_id).await;
 
     let alice_verifies_bob =
         alice.encryption().get_verification(bob.user_id().unwrap(), "flow_id").await;
